Move Gameplay code from GameEngine to Game
