+++ GameEngine/Game.h
@@ -1,21 +1,26 @@
 #pragma once
 #include <SDL.h>
 
+class GameEngine;
+
 class Game
 {
 public:
-	Game() {
-
+	Game(GameEngine* gameEngine) {
+		_gameEngine = gameEngine;
 	}
 
 	void Start();
 
 	void End();
 
-	void Input();
+	void Input(SDL_Event event);
 
 	void Update(Uint32 deltaTime);
 
-	void Render();
+	void Render(SDL_Renderer* renderer);
+
+private:
+	GameEngine* _gameEngine;
 };
